+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -135,6 +135,12 @@ New checks
   Detects default initialization (to 0) of variables with ``enum`` type where
   the enum has no enumerator with value of 0.
 
+- New :doc:`bugprone-null-check-after-dereference
+  <clang-tidy/checks/bugprone/null-check-after-dereference>` check.
+
+  Identifies redundant pointer null-checks, by finding cases where the pointer
+  cannot be null at the location of the null-check.
+
 - New :doc:`llvm-mlir-op-builder
   <clang-tidy/checks/llvm/use-new-mlir-op-builder>` check.